move backup tracker into pkg/backup
